requirements.txt
asgiref
certifi
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.2.2
django-environ==0.10.0
user_agents
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna
Pillow
pycparser
PyJWT
requests
sqlparse
typing-extensions
urllib3
